servers/evolutions-server/pyproject.toml
[project]
name = "evolutions-server"
version = "0.1.0"
description = "MCP Server for evolutions graph data"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp",
    "neo4j>=6.0.2",
    "pyyaml>=6.0.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]
